fix: edit and color code
-modified EDIT url code and some of the color functions
